+++ pybamm/expression_tree/variable.py
@@ -103,8 +103,7 @@ class VariableBase(pybamm.Symbol):
 
     def set_id(self):
         self._id = hash(
-            (self.__class__, self.name, self.scale, self.reference)
-            + tuple([(k, tuple(v)) for k, v in self.domains.items() if v != []])
+            (self.__class__, self.name, self.scale, self.reference, *tuple([(k, tuple(v)) for k, v in self.domains.items() if v != []]))
         )
 
     def create_copy(self):
@@ -129,6 +128,13 @@ class VariableBase(pybamm.Symbol):
             return sympy.Symbol(self.print_name)
         else:
             return self.name
+
+    def to_json(
+        self,
+    ):
+        raise NotImplementedError(
+            "pybamm.Variable: Serialisation is only implemented for discretised models."
+        )
 
 
 class Variable(VariableBase):